requirements.txt
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
Flask-SocketIO==4.3.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
ndg-httpsclient==0.5.1
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-engineio==3.13.2
python-socketio==4.6.0
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0
